Allow Zabbix time spec instead of just seconds for check-delay and disc-delay
In Zabbix, the item check interval and discovery intervals can be
specified not just with a number of seconds, but with more convenient
units using time suffixes as well.

See https://www.zabbix.com/documentation/current/en/manual/appendix/suffixes
for a list of supported suffixes.

By allowing strings as arguments to --check-delay and --disc-delay,
these interval specifications can be used.

No checking for correct syntax (i.e. valid time suffixes) is done; if
required, this can be added to the code as well.